reorganise code and add comments
