+++ frontend/javascripts/logs/log_layer.js
@@ -1,6 +1,7 @@
 /**
- * @namespace javascripts.log.log_layer
- * @description Listen for pushes to window.logLayer and dispatch a loglayerpush event
+ * @namespace javascripts.logs.log_layer
+ * @description Creates a window.logLayer array and dispatches a customer event when set commands are 
+ * performed on the array.
  */
 
 import {logLayerListeners} from "~/javascripts/logs/log_layer_listeners"
@@ -8,6 +9,14 @@ import {logLayerListeners} from "~/javascripts/logs/log_layer_listeners"
 const logLayer = (() => {
   "use strict"
 
+  /**
+   * Create a set trap for the array when data is added or removed. Creates a new custom events which
+   * dispatches the data that has been added. By using Reflect, the original behavior of the push method
+   * is restored while keeping the added behavior.
+   * 
+   * @const _logLayerPushHandler
+   * @memberof javascripts.logs.log_layer
+   */
   const _logLayerPushHandler = {
     set: (target, prop, value) => {
       if (prop !== "length") {
@@ -18,6 +27,14 @@ const logLayer = (() => {
     }
   }
 
+  /**
+   * Create a proxy for a window.logLayer array, which will intercept set commands and introduce
+   * additional functionality in order to handle logging in this application. Additionally sets up
+   * listeners for events dispatched by the proxy.
+   * 
+   * @function init
+   * @memberof javascripts.logs.log_layer
+   */
   const init = () => {
     logLayerListeners.init()
     window.logLayer = new Proxy([], _logLayerPushHandler)
